feat: per-session rooms — decouple Room from Group
A room was pinned to the Group, so every day of a multi-day group was
forced into the same room. Rooms now belong to the session:
GroupSchedule.room, chosen per day, and the conflict check validates that
specific room on that specific day and time.

- Group.room removed; GroupSchedule.room added (nullable, PROTECT).
- ScheduleEntry carries its own room instead of deriving it from the group.
- Group.get_capacity() = smallest room across the group's sessions (a
  student attends every day, so the narrowest room is the real limit).
- Group.get_rooms_display() collapses to one name when all days share a
  room, otherwise lists "day: room".
- Room reverse lookups move to schedules__room / schedule_entries.
- Group create/update and booking screens pick a room per day row;
  the admin's single-day quick form upserts only that day's schedule.
- Session.get_room() resolves the room for that date's weekday.

Migrations are split so 0011 is reversible with no data loss: 0011 adds
GroupSchedule.room and backfills it from each group's current room, 0012
then drops Group.room.

Verified against a restore of production data (47 groups / 87 sessions):
all 87 sessions kept byte-identical room assignments, 81 backfilled and
the 6 that never had a room stayed null. Full suite 610/610.

+++ apps/students/models.py
@@ -247,12 +247,12 @@ class Student(SoftDeleteModel):
         """
         from django.db.models import Max
         from django.db.models.functions import Cast
-        from django.db.models import IntegerField
+        from django.db.models import BigIntegerField
 
         last_numeric = cls.all_objects.filter(
             student_code__regex=r'^\d+$'
         ).annotate(
-            code_int=Cast('student_code', IntegerField())
+            code_int=Cast('student_code', BigIntegerField())
         ).aggregate(max_code=Max('code_int'))
 
         last_code = last_numeric.get('max_code')
